# public/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with TypeScript",
        "version": "1.0.0",
        "description": "A simple CRUD API application made with Express and TypeScript"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "tags": [
        {
            "name": "Listings",
            "description": "API for managing listings"
        }
    ],
    "paths": {
        "/listings": {
            "get": {
                "summary": "Retrieve a list of listings",
                "tags": ["Listings"],
                "responses": {
                    "200": {
                        "description": "A list of listings.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "first listing"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 1000
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "first listing"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new listing",
                "tags": ["Listings"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["id", "description", "price", "title"],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "new listing"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1500
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "new listing title"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Listing created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "new listing"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": 1500
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "new listing title"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/listings/{id}": {
            "delete": {
                "summary": "Delete a listing by ID",
                "tags": ["Listings"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID of the listing to delete.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Listing deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Listing not found"
                    }
                }
            }
        }
    }
}